fix: move article control to admin dashboard
