+++ views/dashboard/articles/create.ejs
@@ -1,10 +1,10 @@
-<% layout('layouts/app') %>
-
-<section class="section-property section-t8">
+<% layout('layouts/dashboard/app') %>
+<!-- Begin Page Content -->
+<div class="container-fluid">
     <div class="row">
-        <h1 class="text-center">Add New Article</h1>
-        <div class="col-6 offset-3">
-            <form action="/articles" method="POST" novalidate enctype="multipart/form-data">
+        <div class="col mx-5 my-3">
+            <h1 class="text-center">Add New Article</h1>
+            <form action="/admin/articles/create" method="POST" novalidate enctype="multipart/form-data">
                 <div class="mb-3">
                     <label for="title" class="form-label">Title</label>
                     <input type="text" name="article[title]" id="title" class="form-control" required />
@@ -28,4 +28,4 @@
             </form>
         </div>
     </div>
-</section>
+</div>
