+++ views/layouts/app.ejs
@@ -23,9 +23,7 @@
 
         <!-- Template Main CSS File -->
         <link href="/css/style.css" rel="stylesheet" />
-        <!-- Tiny -->
-        <script src="https://cdn.tiny.cloud/1/no-api-key/tinymce/5/tinymce.min.js" referrerpolicy="origin"></script>
-        <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+     
     </head>
     <body>
         <%- include('partials/navbar') %> <%- include('partials/alert') %> <%- body %> <%- include('partials/footer') %>
@@ -41,12 +39,7 @@
 
         <!-- Template Main JS File -->
         <script src="/js/main.js"></script>
-        <!-- tiny -->
-        <script>
-            tinymce.init({
-                selector: '#description',
-            })
-        </script>
+       
 
         <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
         <script>
